feat(grok): honor search policy and pin bounded image downloads
