+++ devlog/_plan/260905_production_readiness/055_wp05_build_notes.md
@@ -1,0 +1,49 @@
+# WP05 B — guarded parallel implementation
+
+Real B entered after054A synthesis, docscheckpoint106878d3. Ten bounded workers
+explicitly requested Astra/high; servicepriority not independently set by tool.
+Main owns coordination/FSM/goal/git and exact CI gates. No WP05 PR yet.
+
+| Slice | Worker |
+| --- | --- |
+| Shared DNS/HTTP fixture | Dewey01a070d0-a488-7cd3-aab0-42cf78b90e0d |
+| Planner/facade | Jason01a070d0-a517-7271-a006-31085979d418 |
+| Operations/sequence | Einstein01a070d0-a5d5-70e1-85db-15e5632a3bf1 |
+| Family/legacy | Faraday01a070d0-a67b-71a1-b594-33af9c0d996e |
+| Address policy | Boole01a070d0-a742-7412-a34c-748f9fdcd578 |
+| Download wrapper | Feynman01a070d0-a800-7f53-958c-c5ab730b2509 |
+| Download verification | Peirce01a070d0-a8bd-7d02-86ec-3818df5f7d6e |
+| Runtime Grok cases | Rawls01a070d0-a9d8-77e1-a252-f67a819296cc |
+| Independent legacy fixtures | Arendt01a070d0-aaaa-7251-a476-b1d606c8be25 |
+| Contracts/runner/retry | Epicurus01a070d0-ab62-7f32-9c34-d4a523eda4c1 |
+
+All runtime Grok/legacy tests are STOPPED until main validates the new default-deny
+DNS/HTTP fixture. Source-only AST and genuinely no-network isolated policy/retry/
+tiny canonical-runner probes are allowed. No public DNS/HTTP/provider process or
+paid request; nativepinning only after exact-owned-loopback safety review.
+
+Main implemented genericRetryResponse types with identical algorithms, newfamily
+selector/exclusions, originalIndexes final sweep, canonical modulemock flag and
+CodeQL exact-head dispatch. Missingnewmodule imports mean this is not build-green.
+Main transpile comparison (comments removed) confirms retryJS identical in both
+ESM/CommonJS; this proves the type-only edit, not download runtime acceptance.
+
+Original scripts/recording untouched. RemainingB/C includes strictfixture smoke,
+all planned tests, secret/CodeQL scans, source/HTTP/visual audits, exact-head CI and
+stack publication. No broad completion or release claim.
+
+## Stable implementation handoff
+
+All ten B workers finished. Main reverifiedservertypecheck/build andtesttypecheck0.
+Pure addresspolicy53/53; main fullnetwork+routeharness24/24 afterthebindfix;
+mainruntime4files129/129 (Grok63,node21,multimode6,routes39); legacyplanner/API/
+Agent39/39; contracts/retry/runner63/63, boundary71/71, modeldefaults5/5 and
+prompt-fidelity passed. Counts overlap where suites were rerun, not a unique sum.
+
+Main inspected input AND returned-address guards before running nativepinning:
+ownednamedloopbackHTTP/connect, DNSonce/default0, exactbytes/Host/remoteaddress/
+port/close all passed; two pureguardcases also passed,3/3 total. No publicsocket.
+Small fake-download cases passed perworkers (71/73 subsets); heavy4 mistakenlocal
+execution is disclosed in056, not permission foradditional localstress.
+Allheavy cases still mustpasscanonicalexact-headCI. Source/classifier/mutation/
+CodeQL/gitleaks andmanualQA closure remainC obligations.
