+++ lib/providers/execution/index.ts
@@ -2,6 +2,7 @@ import type { RuntimeContext } from "../../runtimeContext.js";
 import { assertDirectGrokKey } from "./admission.js";
 import { isLegacyExecutionRequest, prepareLegacyImageExecution } from "./legacy.js";
 import { isOpenaiRequest, prepareOpenaiExecution } from "../adapters/openaiExecution.js";
+import { isGrokRequest, prepareGrokExecution } from "../adapters/grokExecution.js";
 import type { ExecutionProgress, ExecutionSurface, ImageExecutionRequest, PreparedImageExecution } from "./types.js";
 
 export type * from "./types.js";
@@ -9,6 +10,7 @@ export type * from "./types.js";
 function prepareSelected(ctx: RuntimeContext, request: ImageExecutionRequest,
   progress?: ExecutionProgress): Promise<PreparedImageExecution<ExecutionSurface>> {
   if (isOpenaiRequest(request)) return prepareOpenaiExecution(ctx, request, progress);
+  if (isGrokRequest(request)) return prepareGrokExecution(ctx, request, progress);
   if (isLegacyExecutionRequest(request)) return prepareLegacyImageExecution(ctx, request, progress);
   throw new Error("Unreachable image execution provider");
 }
