+++ lib/providers/execution/legacy.ts
@@ -6,11 +6,12 @@ import { prepareLegacyEdit } from "./legacyEdit.js";
 import { prepareLegacyMultimode } from "./legacyMultimode.js";
 
 export type LegacyExecutionRequest = ImageExecutionRequest & {
-  provider: Exclude<ImageExecutionRequest["provider"], "oauth" | "api">;
+  provider: Exclude<ImageExecutionRequest["provider"], "oauth" | "api" | "grok" | "grok-api">;
 };
 
 export function isLegacyExecutionRequest(request: ImageExecutionRequest): request is LegacyExecutionRequest {
-  return request.provider !== "oauth" && request.provider !== "api";
+  return request.provider !== "oauth" && request.provider !== "api"
+    && request.provider !== "grok" && request.provider !== "grok-api";
 }
 
 export function prepareLegacyImageExecution<R extends LegacyExecutionRequest>(
